==> Coding Database Package/5.Interest_HK_US_Rate_20250517.ipynb <==
# HK interest rate

import numpy as np
import pandas as pd
import requests
import yfinance as yf

import matplotlib.pyplot as plt

from IPython.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))

apiFrom = '2020-01'
apiTo = '2024-12'

try:
    response = requests.get(f"https://api.hkma.gov.hk/public/market-data-and-statistics/monthly-statistical-bulletin/er-ir/hkd-ir-periodaverage?pagesize=1000&choose=end_of_month&from={apiFrom}&to={apiTo}&sortorder=desc")
    if response.status_code==200 : print('Connect successfully!')
    data = response.json()
except Exception as e:
    print(e)

df_HKi = pd.DataFrame(data['result']['records'])
df_HKi = df_HKi.set_index('end_of_month')  
df_HKi.index = pd.to_datetime(df_HKi.index)
df_HKi = df_HKi.sort_values('end_of_month')
df_HKi.head(2)

df_HKi.info()

filtered_df = df_HKi.savings_deposit_rate[(df_HKi.index.month == 12)]
filtered_df

filtered_df.to_csv('HK_deposit_rate.csv')  

# Federal Funds Effective Rate

# import os
from fredapi import Fred 

# Place your FRED API Key to replace os.environ['FRED_API_KEY']
FRED_API_KEY = '5a0f707c24cb7af846f4dc4b5da84411' # Create free account https://fred.stlouisfed.org/
fred = Fred(api_key=FRED_API_KEY)

apiFrom = '2020-01-01'
apiTo = '2024-12-31'

#df_Fed = pd.DataFrame()
df_Fed = fred.get_series('FEDFUNDS',observation_start = apiFrom, observation_end = apiTo)

pd.DataFrame(data=df_Fed, columns=['Rate'])

df_Fed[(df_Fed.index.month == 12)].to_csv('US_rate.csv')  

# 10 Year Federal Funds Bond Yields

# import os
from fredapi import Fred 

# Place your FRED API Key to replace os.environ['FRED_API_KEY']
FRED_API_KEY = '5a0f707c24cb7af846f4dc4b5da84411' # Create free account https://fred.stlouisfed.org/
fred = Fred(api_key=FRED_API_KEY)

apiFrom = '2020-01-01'
apiTo = '2025-04-30'

#df_Fed = pd.DataFrame()
df_Fed = fred.get_series('IRLTLT01USQ156N',observation_start = apiFrom, observation_end = apiTo)

pd.DataFrame(data=df_Fed)

df_Fed[(df_Fed.index.month == 10)].to_csv('US_yield_rate.csv')
